Reusable onward container trait
