+++ onward/app/controllers/PopularInTag.scala
@@ -1,14 +1,12 @@
 package controllers
 
-import com.gu.facia.client.models.CollectionConfigJson
 import common._
-import layout.{CollectionEssentials, FaciaContainer}
+import containers.Containers
 import model._
 import play.api.mvc.{ RequestHeader, Controller, Action }
 import services._
-import slices.{Fixed, FixedContainers}
 
-object PopularInTag extends Controller with Related with Logging with ExecutionContexts {
+object PopularInTag extends Controller with Related with Containers with Logging with ExecutionContexts {
   def render(tag: String) = Action.async { implicit request =>
     val edition = Edition(request)
     val excludeTags = request.queryString.getOrElse("exclude-tag", Nil)
@@ -19,23 +17,10 @@ object PopularInTag extends Controller with Related with Logging with ExecutionC
   }
 
   private def renderPopularInTag(trails: Seq[Content])(implicit request: RequestHeader) = Cached(600) {
-    val dataId: String = "related content"
-    val displayName = Some(dataId)
-    val properties = FrontProperties.empty
-    val config = CollectionConfigJson.withDefaults(displayName = displayName)
-    val containerDefinition = trails.length match {
-      case 1 => FixedContainers.fixedSmallSlowI
-      case 2 => FixedContainers.fixedSmallSlowII
-      case _ => FixedContainers.fixedMediumFastXII }
 
     val html = views.html.fragments.containers.facia_cards.container(
-      FaciaContainer(
-        1,
-        Fixed(containerDefinition),
-        CollectionConfigWithId(dataId, config),
-        CollectionEssentials(trails take 8, Nil, displayName, None, None, None)
-      ).withTimeStamps,
-      properties
+      onwardContainer("related content", trails),
+      FrontProperties.empty
     )(request)
 
     JsonComponent(
